Refactor Dashboard and Home components for improved user experience and UI consistency
- Simplified the Dashboard component by removing the login functionality and directly checking for wallet information.
- Updated the Home component to include a login form with password handling and error feedback.
- Integrated NextUI components for a modern design in both Dashboard and Home.
- Enhanced user navigation with clear calls to action for creating a new wallet or logging in.
- Removed unused utility functions from the Dashboard component and added them to the Home component for better organization.
